+++ core-java-8/src/main/java/com/baeldung/doublecolon/Computer.java
@@ -1,4 +1,4 @@
-package com.baeldung.doublecolumn;
+package com.baeldung.doublecolon;
 
 public class Computer {
