+++ core-java-8/src/main/java/com/baeldung/doublecolon/function/ComputerPredicate.java
@@ -1,6 +1,6 @@
-package com.baeldung.doublecolumn.function;
+package com.baeldung.doublecolon.function;
 
-import com.baeldung.doublecolumn.Computer;
+import com.baeldung.doublecolon.Computer;
 
 @FunctionalInterface
 public interface ComputerPredicate {
