+++ core-java-8/src/main/java/com/baeldung/doublecolon/function/TriFunction.java
@@ -1,4 +1,4 @@
-package com.baeldung.doublecolumn.function;
+package com.baeldung.doublecolon.function;
 
 import java.util.Objects;
 import java.util.function.Function;
